Added Hashtable With Separate Chaining Implementation
added hashtable with separate chaining implementation, which is implemented with a static array of linked lists(the implementation that exists on this repository). The static array can be resized based on the specified threshold. All necessary methods have been implemeted(AddElement,RemoveElement,GetValue,GetSize,IsEmpty etc). Additionally I added a primitive iterator for my linked list, because I can not be bothered to make the one that c++ supports for the for loop.